test price_link unsubscribe

+++ app/Droit/Inscription/Worker/InscriptionWorker.php
@@ -192,7 +192,9 @@ class InscriptionWorker implements InscriptionWorkerInterface{
 
     public function destroyDocuments($model)
     {
-        \File::delete(collect($model->documents)->pluck('file')->all());
+        if(!empty($model->documents)){
+            \File::delete(collect($model->documents)->pluck('file')->all());
+        }
     }
 
     public function backupDocuments($model)
@@ -202,5 +204,23 @@ class InscriptionWorker implements InscriptionWorkerInterface{
         collect(array_filter($docs))->map(function ($doc, $key) {
             \File::move(public_path($doc), public_path('files/colloques/bak/'.basename($doc)));
         });
+    }
+
+    public function unsubscribe($inscription)
+    {
+        $inscriptions = $inscription->linked_inscriptions ? $inscription->linked_inscriptions : collect([$inscription]);
+
+        foreach ($inscriptions as $simple){
+
+            // Destroy documents or else if we register the same person again the docs are going to be wrong
+            $this->destroyDocuments($simple);
+
+            // If it's a group inscription and we have deleted refresh the groupe invoice and bv
+            if($simple->group_id > 0) {
+                $this->makeDocuments($simple->groupe, true);
+            }
+
+            $this->inscription->delete($simple->id);
+        }
     }
 }
